Some more annoying fixes

+++ src/snakeGym/baseEnv.py
@@ -2,6 +2,7 @@ import socket
 import subprocess
 from datetime import datetime
 from multiprocessing import Process, Queue
+import os
 
 import numpy as np
 
@@ -10,6 +11,8 @@ from .utils import plotLearning
 
 class BaseEnv:
     def __init__(self) -> None:
+        self.action_space = 4
+        self.ManualBattleSnakeCli = "BATTLESNAKE_CLI" in os.environ
         pass
 
     def plotLearning(self, x, scores, epsilons, filename, lines=None):
@@ -31,7 +34,7 @@ class BaseEnv:
                     if x != 0:
                         boardToPrint[yLevel][xLevel] = snakeIcons[snakeNumber]
 
-        for i in boardToPrint:
+        for i in reversed(boardToPrint):
             for j in i:
                 print(j, end="")
             print()
@@ -47,12 +50,13 @@ class BaseEnv:
         for snakeName, snakeUrl in snakes.items():
             snakeUrls += f"--name {snakeName} --url {snakeUrl} "
 
-        self.battleSnakeProc = subprocess.Popen(
-            f"battlesnake play {snakeUrls} -H {self.height} -W {self.width} -g {gamemode} #-t 20000 -o game/{datetime.now().time()}.json".split(),
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-            shell=False,
-        )
+        if not self.ManualBattleSnakeCli:
+            self.battleSnakeProc = subprocess.Popen(
+                f"battlesnake play {snakeUrls} -H {self.height} -W {self.width} -g {gamemode}".split(),
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE,
+                shell=False,
+            )
 
     def startHttpServer(self):
         SERVER_HOST = "127.0.0.1"
@@ -85,7 +89,7 @@ class BaseEnv:
             data = request.splitlines()[-1].encode().decode()
 
             if "move" not in request:
-                client_connection.sendall("HTTP/1.0 200 OK\n".encode())
+                client_connection.sendall("HTTP/1.0 200 OK\n\n{}".encode())
                 client_connection.close()
                 if "end" in request:
                     incomingQueue.put("END")
@@ -105,18 +109,26 @@ class BaseEnv:
 
     def killBattleSnakeRunner(self):
         self.incomingQueue = Queue()
-        self.outgoingQueue = Queue()
+        self.incomingQueue.put(None)
 
-        if self.battleSnakeProc:
-            self.battleSnakeProc.kill()
+        self.outgoingQueue = Queue()
+        self.outgoingQueue.put(None)
+
+        if not self.ManualBattleSnakeCli:
+            if self.battleSnakeProc:
+                self.battleSnakeProc.kill()
 
         if self.server_socket:
             self.server_socket.close()
             self.reader_p.terminate()
             self.reader_p.join()
 
-        while not self.incomingQueue.empty():
-            self.outgoingQueue.get()
+        # print(self.incomingQueue.qsize(), self.outgoingQueue.qsize())
 
-        while not self.incomingQueue.empty():
-            self.outgoingQueue.get()
+        while True:
+            if self.incomingQueue.get() == None:
+                break
+
+        while True:
+            if self.outgoingQueue.get() == None:
+                break
